Simplified checks for metrics func
