+++ metric_noop.go
@@ -1,0 +1,31 @@
+/*
+ * Copyright © 2016-2018 Aeneas Rekkas <aeneas+oss@aeneas.io>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @author		Aeneas Rekkas <aeneas+oss@aeneas.io>
+ * @copyright 	2015-2018 Aeneas Rekkas <aeneas+oss@aeneas.io>
+ * @license 	Apache-2.0
+ */
+
+package ladon
+
+// MetricNoOp is the default metrics implementation , that tracks nothing.
+type MetricNoOp struct{}
+
+func (*MetricNoOp) RequestDeniedBy(r Request, p Policy)                   {}
+func (*MetricNoOp) RequestAllowedBy(r Request, p Policies)                {}
+func (*MetricNoOp) RequestNoMatch(r Request)                              {}
+func (*MetricNoOp) RequestProcessingError(r Request, p Policy, err error) {}
+
+var DefaultMetric = &MetricNoOp{}
